major changes
Added a thus-far untested filtering system to Variant.{cpp,h} which
allows the definition of arbitrary boolean filters for VCF files.  This
will be extended to allow the generation of counts for various filters.
I am developing a command-line interface in the form of vcffilter.cpp.

Added a first pass for vcf statistic generation, vcfstats.cpp.  In its
present form it simply counts a number of important statistics for
individuals in the VCF file and produces a tab-separated table indexed
by sample name.

Now vcflib uses the 'vcf' namespace, so users will have to incorporate
this into any existing applications when upgrading the API.