vcffilter now works (at least for INFO fields)
In practice performance will be poor if you have to parse lots and lots
of sample genotypes for each line, so for large datasets you only want
to filter on INFO fields, use cut.

The use of arbitrary logic filters should allow for complex filtering
patterns.  The implementation of the logic filter uses shunting yard.