update vcfstats to deal with "complex" events
These are alleles which can be interpreted as a block substitution or
multi-base change.  An example would be ATGG -> ACCAG.  The pairwise
alignment of these alleles suggests a mismatch and a gap.

Support is not perfect as pairwise alignment may generate complex
interpretations of some indels.  This should be enough to at least
collect basic statistics about the prevalence of these events in VCF
files.