add vcfroc, vcfcheck, default value for vcf2tsv
vcfroc: Recevier-Operator-Characteristic for VCF files given a truth set
including all true sites and a test file.

vcfcheck: verify that REF fields match reference

Default value for vcf2tsv, allows users to specify a default (e.g. 0 or
NA) for when a field does not have a value in the INFO field of a VCF
record.