edge case fix, numbers as strings in vcffilter
If a number is provided for comparison with a string, the comparison will fail
because it first checks if both the variables being compared have the same
type.  We can deal with this by checking if one is a string and the other a
number at comparison time, and converting both into strings for comparison.
This allows user-expected behavior when doing things like:

    vcffilter -f "Q = 2"

When Q is a string variable in the VCF file.