add direct region specification to vcfintersect
If you have a lot of regions, it may be faster to use tabix or a bed file, but
to quickly specify a target region, such as when filtering VCF files from the
command line, a specific region can be used.  This is in the same numerical
format as BED, but written chrom:start-end.