enable collated merge using "version sort"
One basic problem we encounter with VCF files in the wild is chromosome order.
The natural ordering is best captured by the function filevercmp, from GNU
coreutils, and this functionality is used in vcf-sort --chromosomal-order in
vcflib.  This is now included via a submodule.

vcfcombine provides an important feature which isn't provided via a line-based
sorting method--- samples are correctly ordered in the combined output,
preventing samples swaps.