add method to DAGify a VCF file
Basically, add reference sequence that is not reported in the VCF between
successive records.  Add a uid to each REF and ALT allele, including the
"homozygous" ones added by this process.  These numerical IDs plus the
reference coordinates of the nodes can allow a variant calling processes to
utilize a graph-based mapping.