vcfglbound --- cap GLs and possibly exclude sites with GLs > 0
This normalization process is important when providing GLs as input to phasing
and imputation systems.

Normalization involves scaling all GLs by the max for the sample, so that the
max GL is represented as 0, maintaining the relative values of each GL.  Then,
GLs which are still below a given threshold are set to the threshold to prevent
underflow.  A suitable threshold for typical use is -10, but no defaults are
set in this application.